fix swagger UI not displaying non-CRUD endpoints
adding handlers as lambda expressions does not work as the generated anonymous inner class itself isn't annotated
