+++ cwms-data-api/src/main/java/cwms/cda/api/LevelsController.java
@@ -109,6 +109,7 @@ import org.jooq.DSLContext;
 
 
 public class LevelsController implements CrudHandler {
+    static final String TAG = "Levels";
     private final MetricRegistry metrics;
 
     private final Histogram requestResultSize;
@@ -136,7 +137,7 @@ public class LevelsController implements CrudHandler {
                     required = true),
             method = HttpMethod.POST,
             path = "/levels",
-            tags = {"Levels"}
+            tags = TAG
     )
     @Override
     public void create(@NotNull Context ctx) {
@@ -192,7 +193,7 @@ public class LevelsController implements CrudHandler {
                         + "shall be used."),
             },
             method = HttpMethod.DELETE,
-            tags = {"Levels"})
+            tags = TAG)
     @Override
     public void delete(@NotNull Context ctx, @NotNull String levelId) {
 
@@ -270,7 +271,7 @@ public class LevelsController implements CrudHandler {
                     @OpenApiContent(from = LocationLevels.class, type = Formats.JSONV2)
                 })
             },
-            tags = {"Levels"})
+            tags = TAG)
     @Override
     public void getAll(Context ctx) {
 
@@ -394,7 +395,7 @@ public class LevelsController implements CrudHandler {
                 })
             },
             description = "Retrieves requested Location Level",
-            tags = {"Levels"}
+            tags = TAG
     )
     @Override
     public void getOne(@NotNull Context ctx, @NotNull String levelId) {
@@ -435,7 +436,7 @@ public class LevelsController implements CrudHandler {
                     required = true),
             description = "Update CWMS Location Level",
             method = HttpMethod.PATCH,
-            tags = {"Levels"}
+            tags = TAG
     )
     @Override
     public void update(@NotNull Context ctx, @NotNull String oldLevelId) {
@@ -628,98 +629,5 @@ public class LevelsController implements CrudHandler {
             throw new IllegalArgumentException(e.getMessage());
         }
         return retVal;
-    }
-
-    @OpenApi(
-            pathParams = {
-                @OpenApiParam(name = LEVEL_ID, required = true, description = "Specifies"
-                        + " the requested location level."),
-            },
-            queryParams = {
-                @OpenApiParam(name = OFFICE, required = true, description = "Specifies the "
-                        + "office of the Location Level to be returned"),
-                @OpenApiParam(name = INTERVAL, description = "Interval time step for"
-                        + " the returned time series. Pseudo-regular interval definitions"
-                        + " will be treated like local regular. Irregular interval will generate daily time steps."
-                        + " Default: 0"),
-                @OpenApiParam(name = BEGIN, description = "Specifies the "
-                        + "start of the time window for data to be included in the response. "
-                        + "If this field is not specified, any required time window begins 24"
-                        + " hours prior to the specified or default end time. The format for "
-                        + "this field is ISO 8601 extended, with optional offset and "
-                        + "timezone, i.e., '"
-                        + DATE_FORMAT + "', e.g., '" + EXAMPLE_DATE + "'."),
-                @OpenApiParam(name = END,  description = "Specifies the "
-                        + "end of the time window for data to be included in the response. If"
-                        + " this field is not specified, any required time window ends at the"
-                        + " current time. The format for this field is ISO 8601 extended, "
-                        + "with optional timezone, i.e., '"
-                        + DATE_FORMAT + "', e.g., '" + EXAMPLE_DATE + "'."),
-                @OpenApiParam(name = TIMEZONE,  description = "Specifies "
-                        + "the time zone of the values of the begin and end fields (unless "
-                        + "otherwise specified), as well as the time zone of any times in the"
-                        + " response. If this field is not specified, the default time zone "
-                        + "of UTC shall be used.\r\nIgnored if begin was specified with "
-                        + "offset and timezone."),
-                @OpenApiParam(name = UNIT, required = true, description = "Desired unit for "
-                        + "the values retrieved."),
-            },
-            responses = {
-                @OpenApiResponse(status = STATUS_200,
-                        description = "A CWMS Time Series representation of the specified location level.",
-                        content = {
-                            @OpenApiContent(from = TimeSeries.class, type = Formats.JSONV2),
-                            @OpenApiContent(from = TimeSeries.class, type = Formats.XMLV2),
-                            @OpenApiContent(from = TimeSeries.class, type = Formats.XML),
-                            @OpenApiContent(from = TimeSeries.class, type = Formats.JSON),
-                            @OpenApiContent(from = TimeSeries.class, type = ""),
-                        }),
-                @OpenApiResponse(status = STATUS_400, description = "Invalid parameter combination"),
-                @OpenApiResponse(status = STATUS_404, description = "The provided combination of "
-                        + "parameters did not find a timeseries."),
-                @OpenApiResponse(status = STATUS_501, description = "Requested format is not "
-                        + "implemented")
-            },
-            description = "Retrieves requested Location Level",
-            tags = {"Levels"},
-            method = HttpMethod.GET,
-            path = "/levels/{level-id}/timeseries"
-    )
-    public void getLevelAsTimeSeries(Context ctx) {
-
-        try (final Timer.Context ignored = markAndTime("getLevelAsTimeSeries")) {
-            DSLContext dsl = getDslContext(ctx);
-            Validator<String> pathParam = ctx.pathParamAsClass(LEVEL_ID, String.class);
-            if (!pathParam.hasValue()) {
-                throw new IllegalArgumentException(LEVEL_ID + " path parameter can not be null "
-                        + "when retrieving levels as time series");
-            }
-            String levelId = pathParam.get();
-            String office = requiredParam(ctx, OFFICE);
-            String begin = ctx.queryParam(BEGIN);
-            String end = ctx.queryParam(END);
-            String units = requiredParam(ctx, UNIT);
-            String timezone = ctx.queryParamAsClass(TIMEZONE, String.class).getOrDefault("UTC");
-            String intervalParameter = ctx.queryParamAsClass(INTERVAL, String.class).getOrDefault("0");
-
-            ZoneId tz = ZoneId.of(timezone, ZoneId.SHORT_IDS);
-            begin = begin != null ? begin : "PT-24H";
-
-            ZonedDateTime beginZdt = DateUtils.parseUserDate(begin, timezone);
-            ZonedDateTime endZdt = end != null
-                    ? DateUtils.parseUserDate(end, timezone)
-                    : ZonedDateTime.now(tz);
-
-            LocationLevelsDao levelsDao = getLevelsDao(dsl);
-            Interval interval = IntervalFactory.findAny(IntervalFactory.equalsName(intervalParameter))
-                    .orElseThrow(() -> new IllegalArgumentException("Invalid interval string: "
-                            + intervalParameter + " for location level as timeseries"));
-            JDomLocationLevelRef levelRef = new JDomLocationLevelRef(office, levelId);
-            TimeSeries timeSeries = levelsDao.retrieveLocationLevelAsTimeSeries(levelRef,
-                    beginZdt.toInstant(), endZdt.toInstant(), interval, units);
-            ctx.json(timeSeries);
-            ctx.status(HttpServletResponse.SC_OK);
-        }
-
     }
 }
