+++ cwms-data-api/src/main/java/cwms/cda/api/TimeSeriesController.java
@@ -197,7 +197,7 @@ public class TimeSeriesController implements CrudHandler {
             },
             method = HttpMethod.POST,
             path = "/timeseries",
-            tags = {TAG}
+            tags = TAG
     )
     @Override
     public void create(@NotNull Context ctx) {
@@ -269,7 +269,7 @@ public class TimeSeriesController implements CrudHandler {
         },
         method = HttpMethod.DELETE,
         path = "/timeseries/{timeseries}",
-        tags = {TAG}
+        tags = TAG
     )
     @Override
     public void delete(@NotNull Context ctx, @NotNull String timeseries) {
@@ -403,7 +403,7 @@ public class TimeSeriesController implements CrudHandler {
             },
             method = HttpMethod.GET,
             path = "/timeseries",
-            tags = {TAG}
+            tags = TAG
     )
     @Override
     public void getAll(@NotNull Context ctx) {
@@ -542,7 +542,7 @@ public class TimeSeriesController implements CrudHandler {
             },
             method = HttpMethod.PATCH,
             path = "/timeseries/{timeseries}",
-            tags = {TAG}
+            tags = TAG
     )
     @Override
     public void update(@NotNull Context ctx, @NotNull String id) {
@@ -644,141 +644,5 @@ public class TimeSeriesController implements CrudHandler {
             logger.log(Level.WARNING, null, ex);
         }
         return result.toString();
-    }
-
-    @OpenApi(
-            queryParams = {
-                @OpenApiParam(name = OFFICE, description = "Specifies the owning office of the "
-                        + "timeseries group(s) whose data is to be included in the response. "
-                        + "If this field is not specified, matching timeseries groups "
-                        + "information from all offices shall be returned."),
-                @OpenApiParam(name = CATEGORY_ID, description = "Specifies the category id "
-                        + "of the timeseries to be included in the response.  Optional."),
-                @OpenApiParam(name = GROUP_ID, description = "Specifies the group id "
-                        + "of the timeseries to be included in the response.  Optional."),
-                @OpenApiParam(name = TS_IDS, description = "Specifies a list of timeseries "
-                        + "ids to be included in the response.  Optional. Cannot be used in "
-                        + "combination with category_id and group_id."),
-            },
-            responses = {
-                @OpenApiResponse(status = STATUS_200, content = {
-                    @OpenApiContent(isArray = true, from = Tsv.class, type = Formats.JSON)}),
-                @OpenApiResponse(status = STATUS_404, description = "Based on the combination of "
-                        + "inputs provided the timeseries group(s) were not found."),
-                @OpenApiResponse(status = STATUS_501, description = "request format is not "
-                        + "implemented")
-            },
-            path = "/timeseries/recent",
-            description = "Returns CWMS Timeseries Groups Data",
-            tags = {TAG},
-            method = HttpMethod.GET
-    )
-    public void getRecent(@NotNull Context ctx) {
-
-        try (final Timer.Context ignored = markAndTime("getRecent")) {
-            DSLContext dsl = getDslContext(ctx);
-
-            TimeSeriesDao dao = getTimeSeriesDao(dsl);
-
-            String office = ctx.queryParam(OFFICE);
-            String categoryId = ctx.queryParamAsClass(CATEGORY_ID, String.class).allowNullable().get();
-            String groupId = ctx.pathParamAsClass(GROUP_ID, String.class).allowNullable().get();
-            String tsIdsParam = ctx.queryParamAsClass(TS_IDS, String.class).allowNullable().get();
-
-            GregorianCalendar gregorianCalendar = new GregorianCalendar();
-            gregorianCalendar.set(Calendar.HOUR, 0);
-            gregorianCalendar.set(Calendar.MINUTE, 0);
-            gregorianCalendar.set(Calendar.SECOND, 0);
-            gregorianCalendar.set(Calendar.MILLISECOND, 0);
-
-            gregorianCalendar.add(Calendar.HOUR, 24 * 14);
-            Timestamp futureLimit = Timestamp.from(gregorianCalendar.toInstant());
-            gregorianCalendar.add(Calendar.HOUR, 24 * -28);
-            Timestamp pastLimit = Timestamp.from(gregorianCalendar.toInstant());
-
-            boolean hasTsGroupInfo = categoryId != null && !categoryId.isEmpty()
-                    && groupId != null && !groupId.isEmpty();
-            List<String> tsIds = getTsIds(tsIdsParam);
-            boolean hasTsIds = tsIds != null && !tsIds.isEmpty();
-
-            List<RecentValue> latestValues;
-            if (hasTsGroupInfo && hasTsIds) {
-                // has both = this is an error
-                CdaError re = new CdaError("Invalid arguments supplied, group has both "
-                        + "Timeseries Group info and Timeseries IDs.");
-                logger.log(Level.SEVERE, "{0} for request {1}", new Object[]{ re, ctx.fullUrl()});
-                ctx.status(HttpServletResponse.SC_BAD_REQUEST);
-                ctx.json(re);
-                return;
-            } else if (!hasTsGroupInfo && !hasTsIds) {
-                // doesn't have either?  Just return empty results?
-                CdaError re = new CdaError("Invalid arguments supplied, group has neither "
-                        + "Timeseries Group info nor Timeseries IDs");
-                logger.log(Level.SEVERE, "{0} for request {1}", new Object[]{ re, ctx.fullUrl()});
-                ctx.status(HttpServletResponse.SC_BAD_REQUEST);
-                ctx.json(re);
-                return;
-            } else if (hasTsGroupInfo) {
-                // just group provided
-                latestValues = dao.findRecentsInRange(office, categoryId, groupId, pastLimit,
-                        futureLimit);
-            } else {
-                latestValues = dao.findMostRecentsInRange(tsIds, pastLimit, futureLimit);
-            }
-
-            String formatHeader = ctx.header(Header.ACCEPT);
-            ContentType contentType = Formats.parseHeaderAndQueryParm(formatHeader, null);
-
-            String result = Formats.format(contentType, latestValues, RecentValue.class);
-
-            ctx.result(result).contentType(contentType.toString());
-            requestResultSize.update(result.length());
-
-            ctx.status(HttpServletResponse.SC_OK);
-        }
-    }
-
-    public static List<String> getTsIds(String tsIdsParam) {
-        List<String> retval = null;
-
-        if (tsIdsParam != null && !tsIdsParam.isEmpty()) {
-            retval = new ArrayList<>();
-
-            if (tsIdsParam.startsWith("[")) {
-                tsIdsParam = tsIdsParam.substring(1);
-            }
-
-            if (tsIdsParam.endsWith("]")) {
-                tsIdsParam = tsIdsParam.substring(0, tsIdsParam.length() - 1);
-            }
-
-            if (!tsIdsParam.isEmpty()) {
-                final String regex = "\"[^\"]*\"|[^,]+";
-                final Pattern pattern = Pattern.compile(regex, Pattern.MULTILINE);
-
-                try (Scanner s = new Scanner(tsIdsParam)) {
-                    List<String> matches = findAll(s, pattern).map(m -> m.group().trim())
-                            .collect(Collectors.toList());
-                    retval.addAll(matches);
-                }
-            }
-        }
-
-        return retval;
-    }
-
-    // This came from https://stackoverflow.com/questions/42961296/java-8-stream-emitting-a-stream/42978216#42978216
-    public static Stream<MatchResult> findAll(Scanner s, Pattern pattern) {
-        return StreamSupport.stream(new Spliterators.AbstractSpliterator<MatchResult>(
-                1000, Spliterator.ORDERED | Spliterator.NONNULL) {
-            public boolean tryAdvance(Consumer<? super MatchResult> action) {
-                if (s.findWithinHorizon(pattern, 0) != null) {
-                    action.accept(s.match());
-                    return true;
-                } else {
-                    return false;
-                }
-            }
-        }, false);
     }
 }
